+++ packages/web-client/app/components/workflow-thread/postable/index.hbs
@@ -11,7 +11,7 @@
     @imgURL={{@postable.author.imgURL}}
     @name={{@postable.author.name}}
     @hideName={{true}}
-    @hideMeta={{postable-meta-identical @postable (object-at @postables (dec @index))}}
+    @hideMeta={{postable-meta-hidden @postable previous=(object-at @postables (dec @index))}}
     data-test-postable={{@index}}
     ...attributes
   >
@@ -26,7 +26,7 @@
     @imgURL={{@postable.author.imgURL}}
     @name={{@postable.author.name}}
     @hideName={{true}}
-    @hideMeta={{postable-meta-identical @postable (object-at @postables (dec @index))}}
+    @hideMeta={{postable-meta-hidden @postable previous=(object-at @postables (dec @index))}}
     data-test-postable={{@index}}
     ...attributes
   >
