+++ packages/web-client/app/helpers/postable-meta-identical.ts
@@ -1,13 +1,12 @@
 import { helper } from '@ember/component/helper';
 import { WorkflowPostable } from '../models/workflow/workflow-postable';
-import { WorkflowCard } from '../models/workflow/workflow-card';
 
 // one minute
 const maxIntervalInMilliseconds = 1000 * 60;
 
-function postableMetaIdentical([postA, postB]: [
-  WorkflowPostable | WorkflowCard,
-  WorkflowPostable | WorkflowCard
+export function postableMetaIdentical([postA, postB]: [
+  WorkflowPostable,
+  WorkflowPostable
 ]) {
   if (!postA || !postB) {
     return false;
@@ -18,21 +17,9 @@ function postableMetaIdentical([postA, postB]: [
   let authorA = postA.author.name;
   let authorB = postB.author.name;
 
-  let isSameGroup =
-    Math.abs(timeB - timeA) < maxIntervalInMilliseconds && authorA === authorB;
-
-  if (!isSameGroup) {
-    return false;
-  }
-
-  let postAHasCard = Object.keys(postA).includes('componentName');
-  let postBHasCard = Object.keys(postB).includes('componentName');
-
-  if (postAHasCard || (!postAHasCard && !postBHasCard)) {
-    return true;
-  }
-
-  return false;
+  return (
+    Math.abs(timeB - timeA) < maxIntervalInMilliseconds && authorA === authorB
+  );
 }
 
 export default helper(postableMetaIdentical);
